backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
werkzeug==3.0.1
gunicorn==21.2.0
psutil==5.9.8
psycopg2-binary==2.9.9

